requirements.txt
openai==0.27.5
streamlit==1.22.0
streamlit_authenticator==0.2.2
streamlit_chat @ git+https://github.com/harley/st-chat@main#egg=streamlit_chat
streamlit-oauth
python-dotenv
watchdog
google-cloud-firestore